server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
twikit==2.3.35
anthropic>=0.40.0
python-dotenv==1.0.1
httpx==0.27.2
